+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Added NonNull annotation to FacebookCallback interface
+
 ## [8.1.0] - 2020-10-13
 
 ## Changed